services/assets/pyproject.toml
[project]
name = "spacecases-assets"
version = "0.1.0"
description = "SpaceCases static assets and asset/data generation scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "brotlipy==0.7.0",
    "pillow==11.0.0",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "types-requests==2.32.0.20241016",
    "common"
]

[dependency-groups]
dev = [
    "ruff==0.8.4",
    "pytest==8.3.5",
    "pytest-custom-exit-code==0.3.0",
    "pyright==1.1.396",
]

[tool.uv.workspace]
members = ["../../common"]

[tool.uv.sources]
common = { workspace = true }

[tool.mypy]
exclude = [
    "build",
]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
